// src/types/auth.dto.ts
export type SignInRequestDto = {
  phoneNumber: string;
  password: string;
};

export type AuthResponseDto = {
  access_token: string;
  expires_in: number;
  refresh_expires_in: number;
  refresh_token: string;
  token_type: string;
  'not-before-policy': number;
  session_state: string;
  scope: string;
};